feat(baremetal.commitment): fix issue with upfront price
move logic to main controller to execute on click of next

resolves #DTRSD-56164
