+++ packages/manager/modules/billing-components/src/components/commitment-pricing-mode/commitment-pricing-mode.component.js
@@ -1,12 +1,11 @@
-import controller from './commitment-pricing-mode.controller';
 import template from './commitment-pricing-mode.html';
 
 export default {
   bindings: {
     pricingMode: '=',
     pricingModes: '<',
+    upfrontSavings: '<',
     duration: '<',
   },
-  controller,
   template,
 };
