+++ packages/manager/modules/billing-components/src/components/commitment/commitment.controller.js
@@ -1,4 +1,5 @@
 import { EngagementConfiguration } from '@ovh-ux/manager-models';
+import { convertLanguageFromOVHToBCP47 } from '@ovh-ux/manager-config';
 
 export default class {
   /* @ngInject */
@@ -70,9 +71,9 @@ export default class {
   }
 
   onDurationChange(duration) {
-    this.pricingModes = this.availableEngagements.filter(
-      (commitment) => commitment.durationInMonths === duration.monthlyDuration,
-    );
+    this.pricingModes = this.availableEngagements.filter((commitment) => {
+      return commitment.durationInMonths === duration.monthlyDuration;
+    });
     [this.model.engagement] = this.pricingModes;
   }
 
@@ -98,6 +99,7 @@ export default class {
 
   onPaymentStepFocus() {
     this.isPaymentStepLoading = true;
+    this.getDiscount();
     return this.ovhPaymentMethod
       .getDefaultPaymentMethod()
       .then((paymentMethod) => {
@@ -142,5 +144,49 @@ export default class {
       .catch((error) => {
         this.error = error.data?.message || error.message;
       });
+  }
+
+  static roundToTwo(num) {
+    // https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Number/toFixed
+    return Number.parseFloat(num).toFixed(2);
+  }
+
+  getDiscount() {
+    const upfront = this.pricingModes.find((commitment) =>
+      commitment.isUpfront(),
+    );
+    const periodic = this.pricingModes.find((commitment) =>
+      commitment.isPeriodic(),
+    );
+
+    if (upfront && periodic) {
+      this.discount = this.constructor.roundToTwo(
+        ((periodic.totalPrice.value - upfront.totalPrice.value) * 100) /
+          periodic.totalPrice.value,
+      );
+      this.savings = periodic.getPriceDiff(upfront);
+      let totalSavings = this.savings.value;
+      totalSavings += this.model.duration.savings
+        ? this.model.duration.savings.value
+        : 0;
+      this.upfrontSavings = {
+        amountSaved: this.getPriceAsText(
+          totalSavings,
+          upfront.pricing.price.currencyCode,
+        ),
+        amountToPay: upfront.totalPrice.text,
+      };
+    }
+  }
+
+  getPriceAsText(price, currencyCode) {
+    return Intl.NumberFormat(
+      convertLanguageFromOVHToBCP47(this.coreConfig.getUserLocale()),
+      {
+        style: 'currency',
+        currency: currencyCode,
+        currencyDisplay: 'narrowSymbol',
+      },
+    ).format(price);
   }
 }
