+++ packages/manager/modules/billing-components/src/components/commitment/commitment.html
@@ -65,6 +65,7 @@
         >
             <billing-commitment-pricing-mode
                 ng-if="$ctrl.pricingModes"
+                upfront-savings="$ctrl.upfrontSavings"
                 pricing-mode="$ctrl.model.engagement"
                 pricing-modes="$ctrl.pricingModes"
                 duration="$ctrl.model.duration"
